app/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry
azure-functions==1.23.0
azure-storage-blob==12.26.0
azure-core==1.35.0